+++ lib/dom_event_handler.js
